msm: kgsl: Avoid race of fault handler and recovery
There is a possibility of cx gdsc off timeout when stall on fault
is enabled. smmu fault handler is blocked on device mutex to keep
cx vote and dispatcher trying to turn it off for recovery. To avoid
this race condition, return from fault handler when there is contention
for device mutex.

Change-Id: Ia895152f8cf1ce9bdcfecf256558ba291dabc978
Signed-off-by: gkiranku <gkiranku@codeaurora.org>
Signed-off-by: Chenyang Zhong <zhongcy95@gmail.com>
Signed-off-by: KyuoFoxHuyu <88403766@qq.com>
Signed-off-by: Chuck439 <firmankediri123@gmail.com>